Curate the "real apps" section (#902)
I did some curation work on the "real apps" section. I removed apps which hadn't been updated in a long time, or which didn't have a lot of functionality and so didn't serve as the best kinds of examples. I ended up with a great list of four, which feels just fine.